Input parsing complete
Example: cargo run --bin abrute 4 asdf target

Next is to make it execute decryption attempts in sequence.  Then to
implement threading attempts with multiple cores.  Then to implement a
RAM drive in which to do the attempts in.  Then add log file
integration.  And the last feature to implement is a pooling server and
clients for multiple computers to work on the same task.