Fix zip support. Fixes #15
To get zip unencrypting to work we have to avoid two false positives.
The first is one that returns a successful exit sats one out of every
255 attempts.  The second one is one that dumps empty files out of the
correctly named content.

To get by this I had to implement it so that empty files will not be
kept.  This may not always be desired behavior and perhaps an
additional fix for the situation where some empty files are wanted
should be implemented.